containers/agent-brain/agent-brain/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain>=0.3.0
litellm>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
httpx>=0.27.0
pydantic>=2.8.0
python-dotenv>=1.0.0
playwright>=1.48.0
schedule>=1.2.0
PyJWT>=2.9.0
python-multipart>=0.0.12
aiofiles>=24.1.0
anthropic>=0.34.0